remove e atualiza

+++ produtos-db/src/main/java/fatecfranca/edu/produtos_db/service/ProdutoService.java
@@ -6,6 +6,7 @@ import jakarta.transaction.Transactional;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class ProdutoService {
@@ -20,16 +21,27 @@ public class ProdutoService {
         return repo.findAll();
     }
     // buscar por id
-    public Produto buscar(Long id){
-        return repo.findById(id)
-                .orElseThrow( () -> new RuntimeException("Não encontrado"));
+    public Optional<Produto> buscar(Long id){
+        return repo.findById(id);
     }
     @Transactional
     public Produto salvar(Produto p){
-        return repo.save(p);
+        return repo.save(p); // objeto p sem id
     }
     @Transactional
-    public void remover(Long id){
-        repo.deleteById(id);
+    public boolean remover(Long id){
+        if (repo.existsById(id)){
+            repo.deleteById(id);
+            return true;
+        }
+        return false;
+    }
+    @Transactional
+    public Produto atualizar(Long id, Produto atual){
+        if (repo.existsById(id)){
+            atual.setId(id);
+            return repo.save(atual); // achou
+        }
+        return null; // não achou para atualizar
     }
 }
